public-api/requirements.txt
fastapi
uvicorn
pymongo
passlib
python-consul
httpx
python-jose
